requirements.txt
GitPython
pandas
numpy
scikit-learn
torch
torchtext==0.11.0
transformers
datasets
matplotlib
coloredlogs
jupyter
psutil
humanize
GPUtil
tqdm

# package location
--find-links https://download.pytorch.org/whl/cu113/torch_stable.html

ignite